feat: create EditPresets
