+++ webapp/components/views/Threads/Settings.tsx
@@ -12,30 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 import { useContext } from 'react';
-import { AlertTriangle, Bug, File, HelpCircle, Palette, Settings2, X } from 'lucide-react';
+import { AlertTriangle, Bug, File, Palette, Settings2, X } from 'lucide-react';
 import useTranslation from '@/hooks/useTranslation';
-import logger from '@/utils/logger';
 import { AppContext } from '@/context';
 import useBackend from '@/hooks/useBackendContext';
 import { getConversationAssets, updateConversation } from '@/utils/data/conversations';
 import { findModel } from '@/utils/data/models';
-import Opla from '@/utils/providers/opla';
-import { getCompletionParametersDefinition } from '@/utils/providers';
 import { findProvider, getLocalProvider } from '@/utils/data/providers';
-import { ContextWindowPolicy, Conversation, Preset, PresetParameter } from '@/types';
-import { toast } from '@/components/ui/Toast';
-import { ContextWindowPolicies, DefaultContextWindowPolicy } from '@/utils/constants';
-import { findCompatiblePreset, getCompletePresetProperties } from '@/utils/data/presets';
+import { Conversation, Preset } from '@/types';
 import { getFilename } from '@/utils/misc';
-import { Accordion, AccordionContent, AccordionItem, AccordionTrigger } from '../../ui/accordion';
-import { ScrollArea } from '../../ui/scroll-area';
+import EditPresets from '@/components/common/EditPresets';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '../../ui/tabs';
 import { Textarea } from '../../ui/textarea';
-import Parameter, { ParameterValue, ParametersRecord } from '../../common/Parameter';
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '../../ui/select';
-import { Tooltip, TooltipTrigger, TooltipContent } from '../../ui/tooltip';
-import Form from '../../common/Form';
-import Presets from './Presets';
 import { Button } from '../../ui/button';
 
 export default function Settings({
@@ -46,7 +34,7 @@ export default function Settings({
   errors: string[];
 }) {
   const { t } = useTranslation();
-  const { conversations, updateConversations, providers, presets } = useContext(AppContext);
+  const { conversations, updateConversations, providers } = useContext(AppContext);
   const { backendContext } = useBackend();
 
   const selectedConversation = conversations.find((c) => c.id === conversationId);
@@ -55,88 +43,12 @@ export default function Settings({
   const provider = selectedConversation?.provider
     ? findProvider(selectedConversation?.provider, providers)
     : getLocalProvider(providers);
-  const parametersDefinition = getCompletionParametersDefinition(provider);
-  const modelName = selectedConversation?.model ?? model?.name;
-  const preset = findCompatiblePreset(selectedConversation?.preset, presets, modelName, provider);
-  const {
-    parameters = {},
-    system = selectedConversation?.system ?? model?.system ?? Opla.system,
-    keepSystem,
-    contextWindowPolicy: selectedPolicy = DefaultContextWindowPolicy,
-  } = getCompletePresetProperties(preset, selectedConversation, presets);
 
   const handleNoteChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
     const { value } = e.target;
     if (selectedConversation) {
       const newConversations = updateConversation(
         { ...selectedConversation, note: value },
-        conversations,
-        true,
-      );
-      updateConversations(newConversations);
-    }
-  };
-
-  const handleSystemChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
-    const { value } = e.target;
-    if (selectedConversation) {
-      const newConversations = updateConversation(
-        { ...selectedConversation, system: value },
-        conversations,
-        true,
-      );
-      updateConversations(newConversations);
-    }
-  };
-
-  const updateParameters = async (
-    id: string | undefined,
-    params: ParametersRecord,
-  ): Promise<ParametersRecord | undefined> => {
-    let newParams: ParametersRecord | undefined;
-    if (id && selectedConversation) {
-      let newConversation: Conversation | undefined;
-      newParams = { ...params };
-      let needUpdate = false;
-      Object.keys(params).forEach((key) => {
-        const value = params[key];
-        if (value === undefined) {
-          delete parameters[key];
-          delete newParams?.[key];
-          needUpdate = true;
-        } else {
-          const parameterDef = parametersDefinition[key];
-          const result = parameterDef.z.safeParse(value);
-          if (!result.success) {
-            logger.error('updateParameters invalid', result.error);
-            toast.error(result.error.message);
-          } else {
-            parameters[key] = result.data;
-            delete newParams?.[key];
-            needUpdate = true;
-          }
-        }
-      });
-
-      if (needUpdate) {
-        if (selectedConversation.parameters && Object.keys(parameters).length === 0) {
-          newConversation = { ...selectedConversation };
-          delete newConversation.parameters;
-        } else {
-          newConversation = { ...selectedConversation, parameters };
-        }
-        const newConversations = updateConversation(newConversation, conversations, true);
-        logger.info('onParameterChange save Parameters', params, parameters); // , params);
-        updateConversations(newConversations);
-      }
-    }
-    return newParams;
-  };
-
-  const handlePolicyChange = (policy: ContextWindowPolicy) => {
-    if (selectedConversation) {
-      const newConversations = updateConversation(
-        { ...selectedConversation, contextWindowPolicy: policy },
         conversations,
         true,
       );
@@ -158,22 +70,10 @@ export default function Settings({
     }
   };
 
-  const handleKeepSystemChange = (name: string, value: ParameterValue) => {
+  const handleChangePreset = (newPreset: Conversation) => {
     if (selectedConversation) {
       const newConversations = updateConversation(
-        { ...selectedConversation, keepSystem: value as boolean },
-        conversations,
-        true,
-      );
-
-      updateConversations(newConversations);
-    }
-  };
-
-  const handleChangePreset = (newPreset: string) => {
-    if (selectedConversation) {
-      const newConversations = updateConversation(
-        { ...selectedConversation, preset: newPreset },
+        { ...selectedConversation, ...newPreset },
         conversations,
         true,
       );
@@ -201,75 +101,12 @@ export default function Settings({
           </TabsList>
         </div>
         <TabsContent value="settings" className="h-full py-4">
-          <ScrollArea className="h-full w-full px-4">
-            <Presets
-              preset={preset}
-              presetProperties={selectedConversation as Partial<Preset>}
-              model={modelName}
-              provider={provider}
-              onChangePreset={handleChangePreset}
-            />
-            <Accordion type="multiple" className="w-full px-1" defaultValue={['settings-system']}>
-              <AccordionItem value="settings-system">
-                <AccordionTrigger>{t('System')}</AccordionTrigger>
-                <AccordionContent className="m-0 p-2">
-                  <Textarea
-                    value={system}
-                    onChange={handleSystemChange}
-                    className="min-h-[240px] resize-none  overflow-y-hidden"
-                  />
-                </AccordionContent>
-              </AccordionItem>
-              <AccordionItem value="settings-parameters">
-                <AccordionTrigger>{t('Parameters')}</AccordionTrigger>
-                <AccordionContent>
-                  <Form<PresetParameter>
-                    id={selectedConversation?.id}
-                    parameters={parameters}
-                    parametersDefinition={parametersDefinition}
-                    onParametersChange={updateParameters}
-                  />
-                </AccordionContent>
-              </AccordionItem>
-              <AccordionItem value="contextwindow-parameters">
-                <AccordionTrigger>{t('Context window')}</AccordionTrigger>
-                <AccordionContent className=" my-2 px-2 pb-8">
-                  <div className="flex w-full flex-row py-2">
-                    <Select defaultValue={selectedPolicy} onValueChange={handlePolicyChange}>
-                      <SelectTrigger className="grow capitalize">
-                        <SelectValue placeholder={t('Select policy')} />
-                      </SelectTrigger>
-                      <SelectContent>
-                        {Object.keys(ContextWindowPolicies).map((key) => (
-                          <SelectItem key={key} value={key} className="capitalize">
-                            {t(key)}
-                          </SelectItem>
-                        ))}
-                      </SelectContent>
-                    </Select>
-                    <Tooltip>
-                      <TooltipTrigger className="">
-                        <HelpCircle className="ml-2 h-4 w-4" strokeWidth={1.5} />
-                      </TooltipTrigger>
-                      <TooltipContent side="bottom">
-                        <p className="w-[265px] text-sm">{ContextWindowPolicies[selectedPolicy]}</p>
-                      </TooltipContent>
-                    </Tooltip>
-                  </div>
-
-                  <Parameter
-                    label={t('Keep system')}
-                    type="boolean"
-                    name="keepSystem"
-                    inputCss="max-w-20 pl-2"
-                    value={keepSystem}
-                    description={t('Keep system prompts for the final prompt')}
-                    onChange={handleKeepSystemChange}
-                  />
-                </AccordionContent>
-              </AccordionItem>
-            </Accordion>
-          </ScrollArea>
+          <EditPresets<Conversation>
+            presetProperties={selectedConversation as Partial<Preset>}
+            provider={provider}
+            model={model}
+            onChange={handleChangePreset}
+          />
         </TabsContent>
         <TabsContent value="appearance" className="px-4">
           {t('Thread / Document view')}
